+++ src/matmul_ctx.c
@@ -18,14 +18,14 @@
 /* ----------------------------------------------------------------
  * matmul_ctx_alloc
  * ---------------------------------------------------------------- */
-matmul_ctx_t *matmul_ctx_alloc(size_t dim, const size_t *N, const f64 *L) {
+matmul_ctx_t *matmul_ctx_alloc(uint64_t dim, const uint64_t *N, const f64 *L) {
   matmul_ctx_t *ctx = xcalloc(1, sizeof(matmul_ctx_t));
   ctx->dim = dim;
 
-  ctx->N = xcalloc(dim, sizeof(size_t));
+  ctx->N = xcalloc(dim, sizeof(uint64_t));
   ctx->L = xcalloc(dim, sizeof(f64));
   ctx->size = 1;
-  for (size_t d = 0; d < dim; d++) {
+  for (uint64_t d = 0; d < dim; d++) {
     ctx->N[d] = N[d];
     ctx->L[d] = L[d];
     ctx->size *= N[d];
@@ -54,7 +54,7 @@ void matmul_ctx_free(matmul_ctx_t **pctx) {
 
   /* k-space arrays */
   if (ctx->kx2) {
-    for (size_t d = 0; d < ctx->dim; d++)
+    for (uint64_t d = 0; d < ctx->dim; d++)
       safe_free((void **)&ctx->kx2[d]);
     safe_free((void **)&ctx->kx2);
   }
@@ -95,16 +95,16 @@ void matmul_ctx_set_system(matmul_ctx_t *ctx, int complex_psi0) {
     ctx->k_size = ctx->size;
   } else {
     ctx->k_size = ctx->N[0] / 2 + 1;
-    for (size_t d = 1; d < ctx->dim; d++)
+    for (uint64_t d = 1; d < ctx->dim; d++)
       ctx->k_size *= ctx->N[d];
   }
 
   /* --- Step 2: kx2[d][N[d]] (compact per-dimension) ----------------- */
   ctx->kx2 = xcalloc(ctx->dim, sizeof(f64 *));
-  for (size_t d = 0; d < ctx->dim; d++) {
+  for (uint64_t d = 0; d < ctx->dim; d++) {
     ctx->kx2[d] = xcalloc(ctx->N[d], sizeof(f64));
     const f64 dk = 2.0 * M_PI / ctx->L[d];
-    for (size_t j = 0; j < ctx->N[d]; j++) {
+    for (uint64_t j = 0; j < ctx->N[d]; j++) {
       const f64 kj = (j <= ctx->N[d] / 2)
 	? (dk * (f64)j)
 	: (dk * (f64)((int)j - (int)ctx->N[d]));
@@ -113,20 +113,20 @@ void matmul_ctx_set_system(matmul_ctx_t *ctx, int complex_psi0) {
   }
 
   /* --- Step 3: k2[k_size] (flat, precomputed) ----------------------- */
-  const size_t N0_k = complex_psi0 ? ctx->N[0] : (ctx->N[0] / 2 + 1);
+  const uint64_t N0_k = complex_psi0 ? ctx->N[0] : (ctx->N[0] / 2 + 1);
   ctx->k2 = xcalloc(ctx->k_size, sizeof(f64));
 
   if (1 == ctx->dim) {
-    for (size_t ix = 0; ix < N0_k; ix++)
+    for (uint64_t ix = 0; ix < N0_k; ix++)
       ctx->k2[ix] = ctx->kx2[0][ix];
   } else if (2 == ctx->dim) {
-    for (size_t iy = 0; iy < ctx->N[1]; iy++)
-      for (size_t ix = 0; ix < N0_k; ix++)
+    for (uint64_t iy = 0; iy < ctx->N[1]; iy++)
+      for (uint64_t ix = 0; ix < N0_k; ix++)
 	ctx->k2[iy * N0_k + ix] = ctx->kx2[0][ix] + ctx->kx2[1][iy];
   } else {
-    for (size_t iz = 0; iz < ctx->N[2]; iz++)
-      for (size_t iy = 0; iy < ctx->N[1]; iy++)
-	for (size_t ix = 0; ix < N0_k; ix++)
+    for (uint64_t iz = 0; iz < ctx->N[2]; iz++)
+      for (uint64_t iy = 0; iy < ctx->N[1]; iy++)
+	for (uint64_t ix = 0; ix < N0_k; ix++)
 	  ctx->k2[iz * ctx->N[1] * N0_k + iy * N0_k + ix] =
 	    ctx->kx2[0][ix] + ctx->kx2[1][iy] + ctx->kx2[2][iz];
   }
@@ -140,14 +140,14 @@ void matmul_ctx_set_system(matmul_ctx_t *ctx, int complex_psi0) {
   fftw_plan_with_nthreads(omp_get_max_threads());
 
   /* --- Step 5: Scratch buffers (must precede plan creation) ---------- */
-  const size_t wrk_size = (ctx->size > ctx->k_size) ? ctx->size : ctx->k_size;
+  const uint64_t wrk_size = (ctx->size > ctx->k_size) ? ctx->size : ctx->k_size;
   ctx->f_wrk  = xcalloc(ctx->k_size, sizeof(c64));
   ctx->c_wrk1 = xcalloc(wrk_size, sizeof(c64));
   ctx->c_wrk2 = xcalloc(wrk_size, sizeof(c64));
 
   /* --- Step 6: FFTW plans ------------------------------------------- */
   int fftw_N[3];
-  for (size_t d = 0; d < ctx->dim; d++)
+  for (uint64_t d = 0; d < ctx->dim; d++)
     fftw_N[d] = (int)ctx->N[ctx->dim - 1 - d];
 
   if (complex_psi0) {
